refactor(story-brief): extract generation module
