+++ tests/story_brief/test_weighted_choice.py
@@ -2,10 +2,7 @@ import random
 
 import pytest
 
-from telegraphy.story_brief.generate_story_brief import (
-    symmetric_peak_weights,
-    weighted_choice,
-)
+from telegraphy.story_brief.generation import symmetric_peak_weights, weighted_choice
 
 
 def test_weighted_choice_returns_option_from_domain() -> None:
